refactor: batch of mechanical Sonar cleanups (dead code, redundant casts/types)
Four independent, zero-behavior-change rule violations swept in one pass:
- S1068 unused private fields (ColumnBuilder.dtype x2, PcoAnsEncoder.sizeLog,
  ZipBombSecurityTest.BOMB — the last a stale leftover from before the tests
  switched to small CI-safe row counts).
- S7475 typed unnamed pattern bindings (`var _`/`boolean _` -> bare `_`).
- S1905 redundant numeric casts made no-ops by an already-wide sibling
  operand (binary numeric promotion already widens to long/double there).
- S5838 AssertJ isEqualTo(0) -> isZero() — applied only where the assertion
  target has a concrete numeric static type; reverted on Number-typed
  targets (Compute.filteredSum's return type) where isZero() doesn't exist
  on ObjectAssert<Number> and Sonar's suggestion doesn't actually compile.

+++ writer/src/main/java/io/github/dfa1/vortex/writer/encode/FsstEncodingEncoder.java
@@ -69,12 +69,12 @@ public final class FsstEncodingEncoder implements EncodingEncoder {
     /// @return the terminal `vortex.fsst` encode result
     static EncodeResult toEncodeResult(Fsst c, Arena arena) {
         long uncompLenBytes = c.uncompLenPType().byteSize();
-        MemorySegment uncompLenBuf = arena.allocate(Math.max((long) c.n() * uncompLenBytes, 1));
+        MemorySegment uncompLenBuf = arena.allocate(Math.max(c.n() * uncompLenBytes, 1));
         for (int i = 0; i < c.n(); i++) {
             PTypeIO.set(uncompLenBuf, i * uncompLenBytes, c.uncompLenPType(), c.uncompLens()[i]);
         }
         long codesOffBytes = c.codesOffPType().byteSize();
-        MemorySegment codesOffBuf = arena.allocate((long) (c.n() + 1) * codesOffBytes);
+        MemorySegment codesOffBuf = arena.allocate((c.n() + 1) * codesOffBytes);
         for (int i = 0; i <= c.n(); i++) {
             PTypeIO.set(codesOffBuf, i * codesOffBytes, c.codesOffPType(), c.codesOffsets()[i]);
         }
